fix several small bugs + add timeout to TCP server

+++ server/include/common/utils.h
@@ -4,4 +4,4 @@
 
 ID freshID();
 
-char *readFileBytes(const char *name, int *len);
+char *readFileBytes(const char *name, size_t *len);
